2019.12

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
 
     @PostMapping("/publish")
     public String doPublish(
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
@@ -70,7 +70,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model
     ){
         QuestionDTO question = questionService.getById(id);
